6.5.2

+++ com/baidu/tieba/im/db/e.java
@@ -6,44 +6,45 @@ import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 import android.database.sqlite.SQLiteStatement;
 import android.text.TextUtils;
-import com.baidu.tbadk.TbadkApplication;
+import com.baidu.adp.lib.util.t;
+import com.baidu.tbadk.core.TbadkCoreApplication;
 import com.baidu.tbadk.core.util.TiebaStatic;
 import java.util.Iterator;
 import java.util.LinkedList;
 /* loaded from: classes.dex */
 public class e extends SQLiteOpenHelper {
     public e(Context context, String str) {
-        super(context, str, (SQLiteDatabase.CursorFactory) null, 8);
+        super(context, str, (SQLiteDatabase.CursorFactory) null, 9);
     }
 
-    private void d(SQLiteDatabase sQLiteDatabase, String str) {
+    private void e(SQLiteDatabase sQLiteDatabase, String str) {
         sQLiteDatabase.execSQL(str);
     }
 
     @Override // android.database.sqlite.SQLiteOpenHelper
     public void onCreate(SQLiteDatabase sQLiteDatabase) {
-        n(sQLiteDatabase);
+        q(sQLiteDatabase);
     }
 
-    private void n(SQLiteDatabase sQLiteDatabase) {
+    private void q(SQLiteDatabase sQLiteDatabase) {
         try {
-            d(sQLiteDatabase, "CREATE TABLE IF NOT EXISTS tb_group_news(notice_id TEXT NOT NULL UNIQUE, cmd TEXT, gid TEXT, time long, content TEXT, content_status int, ext TEXT);");
-            o(sQLiteDatabase);
-            d(sQLiteDatabase, "CREATE TABLE IF NOT EXISTS tb_new_friends(_id INTEGER primary key autoincrement, uid INTEGER NOT NULL, uname TEXT, ustatus INTEGER, uportrait TEXT, ucontent TEXT,isread INTEGER);");
+            e(sQLiteDatabase, "CREATE TABLE IF NOT EXISTS tb_group_news(notice_id TEXT NOT NULL UNIQUE, cmd TEXT, gid TEXT, time long, content TEXT, content_status int, ext TEXT);");
+            r(sQLiteDatabase);
+            e(sQLiteDatabase, "CREATE TABLE IF NOT EXISTS tb_new_friends(_id INTEGER primary key autoincrement, uid INTEGER NOT NULL, uname TEXT, ustatus INTEGER, uportrait TEXT, ucontent TEXT,isread INTEGER);");
         } catch (Exception e) {
             TiebaStatic.printDBExceptionLog(e, "ImDatabaseHelper.createTables", new Object[0]);
         }
     }
 
-    private void o(SQLiteDatabase sQLiteDatabase) {
+    private void r(SQLiteDatabase sQLiteDatabase) {
         if (sQLiteDatabase != null) {
-            d(sQLiteDatabase, "CREATE TABLE IF NOT EXISTS tb_message_center(gid TEXT NOT NULL, group_name TEXT, group_head TEXT, custom_group_type int, group_ext TEXT, unread_count int, last_msgId TEXT, last_user_name TEXT, last_content_time long, send_status int, last_content TEXT, is_friend int, pull_msgid TEXT,is_hidden int,is_delete int, sent_mid long, read_mid long);");
+            e(sQLiteDatabase, "CREATE TABLE IF NOT EXISTS tb_message_center(gid TEXT NOT NULL, group_name TEXT, group_head TEXT, custom_group_type int, group_ext TEXT, unread_count int, last_msgId TEXT, last_user_name TEXT, last_content_time long, send_status int, last_content TEXT, is_friend int, pull_msgid TEXT,is_hidden int,is_delete int, sent_mid long, read_mid long, user_type int default 1);");
         }
     }
 
-    private void p(SQLiteDatabase sQLiteDatabase) {
+    private void s(SQLiteDatabase sQLiteDatabase) {
         if (sQLiteDatabase != null) {
-            d(sQLiteDatabase, "CREATE TABLE IF NOT EXISTS tb_message_center(gid TEXT NOT NULL, group_name TEXT, group_head TEXT, custom_group_type int, group_ext TEXT, unread_count int, last_msgId TEXT, last_user_name TEXT, last_content_time long, send_status int, last_content TEXT, is_friend int, pull_msgid TEXT,is_hidden int,is_delete int);");
+            e(sQLiteDatabase, "CREATE TABLE IF NOT EXISTS tb_message_center(gid TEXT NOT NULL, group_name TEXT, group_head TEXT, custom_group_type int, group_ext TEXT, unread_count int, last_msgId TEXT, last_user_name TEXT, last_content_time long, send_status int, last_content TEXT, is_friend int, pull_msgid TEXT,is_hidden int,is_delete int);");
         }
     }
 
@@ -52,64 +53,74 @@ public class e extends SQLiteOpenHelper {
         switch (i) {
             case 1:
                 try {
-                    v(sQLiteDatabase);
-                    q(sQLiteDatabase);
-                    r(sQLiteDatabase);
-                    s(sQLiteDatabase);
+                    y(sQLiteDatabase);
                     t(sQLiteDatabase);
+                    u(sQLiteDatabase);
+                    v(sQLiteDatabase);
                     w(sQLiteDatabase);
-                    x(sQLiteDatabase);
+                    z(sQLiteDatabase);
+                    A(sQLiteDatabase);
+                    B(sQLiteDatabase);
                     return;
                 } catch (Exception e) {
                     TiebaStatic.printDBExceptionLog(e, "ImDatabaseHelper.onUpgrade", new Object[0]);
-                    TbadkApplication.m251getInst().getApp().deleteDatabase(String.valueOf(TbadkApplication.getCurrentAccount()) + ".db");
-                    n(sQLiteDatabase);
+                    TbadkCoreApplication.m255getInst().getApp().deleteDatabase(String.valueOf(TbadkCoreApplication.getCurrentAccount()) + ".db");
+                    q(sQLiteDatabase);
                     return;
                 }
             case 2:
-                q(sQLiteDatabase);
-                r(sQLiteDatabase);
-                s(sQLiteDatabase);
                 t(sQLiteDatabase);
+                u(sQLiteDatabase);
+                v(sQLiteDatabase);
                 w(sQLiteDatabase);
-                x(sQLiteDatabase);
+                z(sQLiteDatabase);
+                A(sQLiteDatabase);
+                B(sQLiteDatabase);
                 return;
             case 3:
-                r(sQLiteDatabase);
-                s(sQLiteDatabase);
-                t(sQLiteDatabase);
+                u(sQLiteDatabase);
+                v(sQLiteDatabase);
                 w(sQLiteDatabase);
-                x(sQLiteDatabase);
+                z(sQLiteDatabase);
+                A(sQLiteDatabase);
+                B(sQLiteDatabase);
                 return;
             case 4:
-                s(sQLiteDatabase);
-                t(sQLiteDatabase);
+                v(sQLiteDatabase);
                 w(sQLiteDatabase);
-                x(sQLiteDatabase);
+                z(sQLiteDatabase);
+                A(sQLiteDatabase);
+                B(sQLiteDatabase);
                 return;
             case 5:
-                t(sQLiteDatabase);
                 w(sQLiteDatabase);
-                x(sQLiteDatabase);
+                z(sQLiteDatabase);
+                A(sQLiteDatabase);
+                B(sQLiteDatabase);
                 return;
             case 6:
-                w(sQLiteDatabase);
-                x(sQLiteDatabase);
+                z(sQLiteDatabase);
+                A(sQLiteDatabase);
+                B(sQLiteDatabase);
                 return;
             case 7:
-                x(sQLiteDatabase);
+                A(sQLiteDatabase);
+                B(sQLiteDatabase);
+                return;
+            case 8:
+                B(sQLiteDatabase);
                 return;
             default:
                 return;
         }
     }
 
-    private void q(SQLiteDatabase sQLiteDatabase) {
+    private void t(SQLiteDatabase sQLiteDatabase) {
         if (sQLiteDatabase != null) {
-            LinkedList<String> u = u(sQLiteDatabase);
+            LinkedList<String> x = x(sQLiteDatabase);
             try {
                 sQLiteDatabase.beginTransaction();
-                Iterator<String> it = u.iterator();
+                Iterator<String> it = x.iterator();
                 while (it.hasNext()) {
                     String next = it.next();
                     if (!TextUtils.isEmpty(next) && next.startsWith("tb_group_msg_")) {
@@ -131,10 +142,10 @@ public class e extends SQLiteOpenHelper {
         }
     }
 
-    private void r(SQLiteDatabase sQLiteDatabase) {
+    private void u(SQLiteDatabase sQLiteDatabase) {
     }
 
-    private void s(SQLiteDatabase sQLiteDatabase) {
+    private void v(SQLiteDatabase sQLiteDatabase) {
         if (sQLiteDatabase != null) {
             try {
                 sQLiteDatabase.beginTransaction();
@@ -151,16 +162,16 @@ public class e extends SQLiteOpenHelper {
         }
     }
 
-    private void t(SQLiteDatabase sQLiteDatabase) {
-        d(sQLiteDatabase, "CREATE TABLE IF NOT EXISTS tb_new_friends(_id INTEGER primary key autoincrement, uid INTEGER NOT NULL, uname TEXT, ustatus INTEGER, uportrait TEXT, ucontent TEXT,isread INTEGER);");
+    private void w(SQLiteDatabase sQLiteDatabase) {
+        e(sQLiteDatabase, "CREATE TABLE IF NOT EXISTS tb_new_friends(_id INTEGER primary key autoincrement, uid INTEGER NOT NULL, uname TEXT, ustatus INTEGER, uportrait TEXT, ucontent TEXT,isread INTEGER);");
         if (sQLiteDatabase != null) {
-            LinkedList<String> u = u(sQLiteDatabase);
+            LinkedList<String> x = x(sQLiteDatabase);
             try {
                 sQLiteDatabase.beginTransaction();
-                Iterator<String> it = u.iterator();
+                Iterator<String> it = x.iterator();
                 while (it.hasNext()) {
                     String next = it.next();
-                    if (!TextUtils.isEmpty(next) && (next.startsWith(o.aVl) || next.startsWith(n.aVl))) {
+                    if (!TextUtils.isEmpty(next) && (next.startsWith(o.aWD) || next.startsWith(n.aWD))) {
                         try {
                             sQLiteDatabase.execSQL("ALTER TABLE " + next + " ADD is_friend int default 1;");
                         } catch (Exception e) {
@@ -176,7 +187,7 @@ public class e extends SQLiteOpenHelper {
         }
     }
 
-    private LinkedList<String> u(SQLiteDatabase sQLiteDatabase) {
+    private LinkedList<String> x(SQLiteDatabase sQLiteDatabase) {
         Cursor cursor;
         Throwable th;
         Exception exc;
@@ -195,17 +206,17 @@ public class e extends SQLiteOpenHelper {
                             try {
                                 TiebaStatic.printDBExceptionLog(exc, "ImDatabaseHelper.getAllTable", new Object[0]);
                                 exc.printStackTrace();
-                                com.baidu.adp.lib.util.p.a(cursor);
+                                t.b(cursor);
                                 return linkedList;
                             } catch (Throwable th2) {
                                 th = th2;
-                                com.baidu.adp.lib.util.p.a(cursor);
+                                t.b(cursor);
                                 throw th;
                             }
                         } catch (Throwable th3) {
                             cursor = cursor2;
                             th = th3;
-                            com.baidu.adp.lib.util.p.a(cursor);
+                            t.b(cursor);
                             throw th;
                         }
                     }
@@ -218,16 +229,16 @@ public class e extends SQLiteOpenHelper {
                 th = th4;
             }
         }
-        com.baidu.adp.lib.util.p.a(cursor2);
+        t.b(cursor2);
         return linkedList;
     }
 
-    private void v(SQLiteDatabase sQLiteDatabase) {
+    private void y(SQLiteDatabase sQLiteDatabase) {
         if (sQLiteDatabase != null) {
-            LinkedList<String> u = u(sQLiteDatabase);
+            LinkedList<String> x = x(sQLiteDatabase);
             try {
                 sQLiteDatabase.beginTransaction();
-                Iterator<String> it = u.iterator();
+                Iterator<String> it = x.iterator();
                 while (it.hasNext()) {
                     String next = it.next();
                     if (!TextUtils.isEmpty(next) && next.startsWith("tb_group_msg_")) {
@@ -246,29 +257,29 @@ public class e extends SQLiteOpenHelper {
         }
     }
 
-    private void w(SQLiteDatabase sQLiteDatabase) {
+    private void z(SQLiteDatabase sQLiteDatabase) {
         if (sQLiteDatabase != null) {
             try {
                 sQLiteDatabase.beginTransaction();
                 try {
                     sQLiteDatabase.execSQL("ALTER TABLE tb_message_center RENAME TO test_table_temp;");
-                    p(sQLiteDatabase);
+                    s(sQLiteDatabase);
                     sQLiteDatabase.execSQL("INSERT INTO tb_message_center SELECT gid,group_name,group_head,custom_group_type,group_ext,unread_count,last_msgId,last_user_name,last_content_time,'send_status' as send_status,last_content,'is_friend' as is_friend,'pull_msgid' as pull_msgid,is_hidden,is_delete FROM test_table_temp;");
                     sQLiteDatabase.execSQL("DROP TABLE test_table_temp;");
                 } catch (Exception e) {
                     TiebaStatic.printDBExceptionLog(e, "ImDatabaseHelper.messageCenter6to7", new Object[0]);
                     sQLiteDatabase.execSQL("DROP TABLE IF EXISTS tb_message_center");
-                    n(sQLiteDatabase);
+                    q(sQLiteDatabase);
                 }
                 sQLiteDatabase.setTransactionSuccessful();
                 try {
                     sQLiteDatabase.beginTransaction();
-                    f.aVr.put(String.valueOf(TbadkApplication.getCurrentAccount()) + ".db", sQLiteDatabase);
-                    l.MN();
+                    f.aWJ.put(String.valueOf(TbadkCoreApplication.getCurrentAccount()) + ".db", sQLiteDatabase);
+                    l.MH();
                     sQLiteDatabase.setTransactionSuccessful();
                 } finally {
                     sQLiteDatabase.endTransaction();
-                    f.aVr.remove(String.valueOf(TbadkApplication.getCurrentAccount()) + ".db");
+                    f.aWJ.remove(String.valueOf(TbadkCoreApplication.getCurrentAccount()) + ".db");
                 }
             } finally {
                 sQLiteDatabase.endTransaction();
@@ -276,7 +287,7 @@ public class e extends SQLiteOpenHelper {
         }
     }
 
-    private void x(SQLiteDatabase sQLiteDatabase) {
+    private void A(SQLiteDatabase sQLiteDatabase) {
         if (sQLiteDatabase != null) {
             try {
                 sQLiteDatabase.beginTransaction();
@@ -285,6 +296,23 @@ public class e extends SQLiteOpenHelper {
                     sQLiteDatabase.execSQL("ALTER TABLE tb_message_center ADD read_mid long default 0;");
                 } catch (Exception e) {
                     TiebaStatic.printDBExceptionLog(e, "ImDatabaseHelper.groupMsg7to8", new Object[0]);
+                    sQLiteDatabase.execSQL("DROP TABLE IF EXISTS tb_message_center");
+                }
+                sQLiteDatabase.setTransactionSuccessful();
+            } finally {
+                sQLiteDatabase.endTransaction();
+            }
+        }
+    }
+
+    private void B(SQLiteDatabase sQLiteDatabase) {
+        if (sQLiteDatabase != null) {
+            try {
+                sQLiteDatabase.beginTransaction();
+                try {
+                    sQLiteDatabase.execSQL("ALTER TABLE tb_message_center ADD user_type int default  1;");
+                } catch (Exception e) {
+                    TiebaStatic.printDBExceptionLog(e, "ImDatabaseHelper.groupMsg8to9", new Object[0]);
                     sQLiteDatabase.execSQL("DROP TABLE IF EXISTS tb_message_center");
                 }
                 sQLiteDatabase.setTransactionSuccessful();
